+++ cdm_reader_mapper/cdm_mapper/__init__.py
@@ -5,6 +5,5 @@ from __future__ import annotations
 from .duplicates import duplicate_check  # noqa
 from .mapper import map_model  # noqa
 from .table_reader import read_tables  # noqa
-from .table_writer import cdm_to_ascii  # noqa
-from .table_writer import table_to_ascii  # noqa
+from .table_writer import write_tables  # noqa
 from .tables.tables import get_cdm_atts  # noqa